requirements.txt
requests
python-daemon
pyyaml
websocket-client
slackclient
Google-Search-API
Pillow
imgurpython
